Module wasmer_package::convert

Modules

error 🔒

webc_to_package 🔒

Structs

ConversionError

Functions

webc_to_package_dir

Convert a webc image into a directory with a wasmer.toml file that can be used for generating a new pacakge.