wasmer_package::convert

Function webc_to_package_dir

pub fn webc_to_package_dir(
    webc: &Container,
    target_dir: &Path,
) -> Result<(), ConversionError>

Convert a webc image into a directory with a wasmer.toml file that can be used for generating a new package.