Module wasmer_package::package

Load a Wasmer package from disk.

Modules

manifest 🔒

package 🔒

strictness 🔒

volume 🔒

Structs

FsVolume

A lazily loaded volume in a Wasmer package.

MemoryDir

An in-memory directory.

MemoryFile

An in-memory file.

MemoryVolume

An in-memory representation of a volume.

Package

A Wasmer package that will be lazily loaded from disk.

Enums

ManifestError

Errors that may occur when converting from a wasmer_config::package::Manifest to a [crate::metadata::Manifest].

MemoryNode

An in-memory representation of a filesystem node.

Strictness

The strictness to use when working with a [crate::wasmer_package::Package].

WasmerPackageError

Errors that may occur while loading a Wasmer package from disk.

Traits

WasmerPackageVolume

An abstraction over concrete volumes implementation to be used in a Wasmer Package.