Enum wasmer_package::package::manifest::ManifestError

#[non_exhaustive]
pub enum ManifestError {

    InvalidDependency(String),
    SerializeCborAnnotation {
        key: String,
        error: Error,
    },
    UnknownAtomKind(String),
    DuplicateModule(String),
    ReadAtomFile {
        module: String,
        path: PathBuf,
        error: Error,
    },
    DuplicateCommand(String),
    UnknownRunnerKind(String),
    MergeAnnotations {
        command: String,
        key: String,
    },
    InvalidModuleReference {
        command: String,
        module: String,
    },
    UndeclaredCommandDependency {
        command: String,
        dependency: String,
    },
    WasmerTomlAnnotations {
        error: Box<dyn Error + Send + Sync>,
    },
    OutsideBaseDirectory {
        path: PathBuf,
        base_dir: PathBuf,
    },
    MissingFile {
        path: PathBuf,
        base_dir: PathBuf,
    },
    FileNotSupported,

}

Errors that may occur when converting from a wasmer_config::package::Manifest to a [crate::metadata::Manifest].

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

InvalidDependency(String)

A dependency specification had a syntax error.

SerializeCborAnnotation

Unable to serialize an annotation.

Fields

key: String

Which annotation was being serialized?

error: Error

The underlying error.

UnknownAtomKind(String)

Specified an unknown atom kind.

DuplicateModule(String)

A module was specified more than once.

ReadAtomFile

Unable to read a moduleโs source.

Fields

module: String

The name of the module.

path: PathBuf

The path that was read.

error: Error

The underlying error.

DuplicateCommand(String)

A command was specified more than once.

UnknownRunnerKind(String)

An unknown runner kind was specified.

MergeAnnotations

An error occurred while merging user-defined annotations in with automatically generated ones.

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

command: String

The command annotations were being merged for.

key: String

The annotation that was being merged.

InvalidModuleReference

A command uses a non-existent module.

Fields

command: String

The command name.

module: String

The module name.

UndeclaredCommandDependency

A command references a module from an undeclared dependency.

Fields

command: String

The command name.

dependency: String

The dependency name.

WasmerTomlAnnotations

Unable to deserialize custom annotations from the wasmer.toml manifest.

Fields

error: Box<dyn Error + Send + Sync>

The underlying error.

OutsideBaseDirectory

The wasmer.toml file references a file outside of its base directory.

Fields

path: PathBuf

The file that was referenced.

base_dir: PathBuf

The base directory.

MissingFile

The manifest references a file that doesnโt exist.

Fields

path: PathBuf

The file that was referenced.

base_dir: PathBuf

The base directory.

FileNotSupported

File based commands are not supported for in-memory package creation