wasmer_package::package
Module package

Structs

Package

A Wasmer package that will be lazily loaded from disk.

Enums

BaseDir 🔒

WasmerPackageError

Errors that may occur while loading a Wasmer package from disk.

Constants

IS_WASI 🔒

Functions

read_manifest 🔒

tempdir 🔒

A polyfill for TempDir::new() that will work when compiling to WASI-based targets.

unpack_archive 🔒

A polyfill for Archive::unpack() that is WASI-compatible.