wasmer_package::package::package

Enum WasmerPackageError

#[non_exhaustive]
pub enum WasmerPackageError {

    TempDir(Error),
    FileOpen {
        path: PathBuf,
        error: Error,
    },
    FileRead {
        path: PathBuf,
        error: Error,
    },
    IoError(Error),
    MalformedPath(PathBuf),
    Tarball(Error),
    TomlDeserialize {
        path: PathBuf,
        error: Error,
    },
    JsonDeserialize {
        path: PathBuf,
        error: Error,
    },
    MissingManifest,
    Canonicalize {
        path: PathBuf,
        error: Error,
    },
    Manifest(ManifestError),
    Validation(ValidationError),
    PathNotExists {
        path: PathBuf,
    },
    VolumeCreation(Error),
    SerdeError(Error),
    ContainerError(ContainerError),
    DetectError(DetectError),

}

Errors that may occur while loading a Wasmer package from disk.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

TempDir(Error)

Unable to create a temporary directory.

FileOpen

Unable to open a file.

Fields

path: PathBuf

The file being opened.

error: Error

The underlying error.

FileRead

Unable to read a file.

Fields

path: PathBuf

The file being opened.

error: Error

The underlying error.

IoError(Error)

Generic IO error.

MalformedPath(PathBuf)

Unexpected path format

Tarball(Error)

Unable to extract the tarball.

TomlDeserialize

Unable to deserialize the wasmer.toml file.

Fields

path: PathBuf

The file being deserialized.

error: Error

The underlying error.

JsonDeserialize

Unable to deserialize a json file.

Fields

path: PathBuf

The file being deserialized.

error: Error

The underlying error.

MissingManifest

Unable to find the wasmer.toml file.

Canonicalize

Unable to canonicalize a path.

Fields

path: PathBuf

The path being canonicalized.

error: Error

The underlying error.

Manifest(ManifestError)

Unable to load the wasmer.toml manifest.

Validation(ValidationError)

A manifest validation error.

PathNotExists

A path in the fs mapping does not exist

Fields

path: PathBuf

Path entry in fs mapping

VolumeCreation(Error)

Any error happening when populating the volumes tree map of a package

SerdeError(Error)

Error when serializing or deserializing

ContainerError(ContainerError)

Container Error

DetectError(DetectError)

Detect Error