Struct wasmer_package::package::package::Package

pub struct Package {
    base_dir: BaseDir,
    manifest: Manifest,
    atoms: BTreeMap<String, OwnedBuffer>,
    strictness: Strictness,
    volumes: BTreeMap<String, Arc<dyn WasmerPackageVolume + Send + Sync + 'static>>,
}

A Wasmer package that will be lazily loaded from disk.

Fields
base_dir: BaseDirmanifest: Manifestatoms: BTreeMap<String, OwnedBuffer>strictness: Strictnessvolumes: BTreeMap<String, Arc<dyn WasmerPackageVolume + Send + Sync + 'static>>
Implementations

impl Package

pub fn from_tarball_file( path: impl AsRef<Path>, ) -> Result<Self, WasmerPackageError>

Load a Package from a *.tar.gz file on disk.

Implementation Details

This will unpack the tarball to a temporary directory on disk and use memory-mapped files in order to reduce RAM usage.

pub fn from_tarball_file_with_strictness( path: impl AsRef<Path>, strictness: Strictness, ) -> Result<Self, WasmerPackageError>

Load a Package from a *.tar.gz file on disk.

Implementation Details

This will unpack the tarball to a temporary directory on disk and use memory-mapped files in order to reduce RAM usage.

pub fn from_tarball(tarball: impl BufRead) -> Result<Self, WasmerPackageError>

Load a package from a *.tar.gz archive.

pub fn from_tarball_with_strictness( tarball: impl BufRead, strictness: Strictness, ) -> Result<Self, WasmerPackageError>

Load a package from a *.tar.gz archive.

pub fn from_manifest( wasmer_toml: impl AsRef<Path>, ) -> Result<Self, WasmerPackageError>

Load a package from a wasmer.toml manifest on disk.

pub fn from_manifest_with_strictness( wasmer_toml: impl AsRef<Path>, strictness: Strictness, ) -> Result<Self, WasmerPackageError>

Load a package from a wasmer.toml manifest on disk.

pub fn from_json_manifest(manifest: PathBuf) -> Result<Self, WasmerPackageError>

(Re)loads a package from a manifest.json file which was created as the result of calling Container::unpack

pub fn from_json_manifest_with_strictness( manifest: PathBuf, strictness: Strictness, ) -> Result<Self, WasmerPackageError>

(Re)loads a package from a manifest.json file which was created as the result of calling Container::unpack

pub fn from_in_memory( manifest: WasmerManifest, volumes: BTreeMap<String, MemoryVolume>, atoms: BTreeMap<String, (Option<String>, OwnedBuffer)>, metadata: MemoryVolume, strictness: Strictness, ) -> Result<Self, WasmerPackageError>

Create a Package from an in-memory representation.

fn load( wasmer_toml: WasmerManifest, base_dir: impl Into<BaseDir>, strictness: Strictness, ) -> Result<Self, WasmerPackageError>

pub fn webc_hash(&self) -> Option<[u8; 32]>

Returns the Sha256 has of the webc represented by this Package

pub fn manifest(&self) -> &WebcManifest

Get the WEBC manifest.

pub fn atoms(&self) -> &BTreeMap<String, OwnedBuffer>

Get all atoms in this package.

pub fn volumes( &self, ) -> impl Iterator<Item = &Arc<dyn WasmerPackageVolume + Sync + Send + 'static>>

Returns all volumes in this package

pub fn serialize(&self) -> Result<Bytes, Error>

Serialize the package to a *.webc v2 file, ignoring errors due to missing files.

fn atom_entries(&self) -> Result<BTreeMap<PathSegment, FileEntry<'_>>, Error>

pub(crate) fn get_volume( &self, name: &str, ) -> Option<Arc<dyn WasmerPackageVolume + Sync + Send + 'static>>

pub(crate) fn volume_names(&self) -> Vec<Cow<'_, str>>