Struct wasmer_package::package::FsVolume

pub struct FsVolume {
    name: String,
    intermediate_directories: BTreeSet<PathBuf>,
    metadata_files: BTreeSet<PathBuf>,
    mapped_directories: BTreeSet<PathBuf>,
    base_dir: PathBuf,
}

A lazily loaded volume in a Wasmer package.

Note that it is the package resolverโs role to interpret a packageโs [crate::metadata::annotations::FileSystemMappings]. A [Volume] contains directories as they were when the package was published.

Fields
name: String

Name of the volume

intermediate_directories: BTreeSet<PathBuf>

A pre-computed set of intermediate directories that are needed to allow access to the whitelisted files and directories.

metadata_files: BTreeSet<PathBuf>

Specific files that this volume has access to.

mapped_directories: BTreeSet<PathBuf>

Directories that allow the user to access anything inside them.

base_dir: PathBuf

The base directory all [PathSegments] will be resolved relative to.

Implementations

impl FsVolume

pub(crate) const METADATA: &'static str = "metadata"

The name of the volume used to store metadata files.

pub(crate) fn new_metadata( manifest: &Manifest, base_dir: impl Into<PathBuf>, ) -> Result<Self, Error>

Create a new metadata volume.

pub(crate) fn new_assets( manifest: &Manifest, base_dir: &Path, ) -> Result<BTreeMap<String, Self>, Error>

pub(crate) fn new_with_intermediate_dirs( name: String, base_dir: PathBuf, whitelisted_files: BTreeSet<PathBuf>, whitelisted_directories: BTreeSet<PathBuf>, ) -> Self

pub(crate) fn new( name: String, base_dir: PathBuf, whitelisted_files: BTreeSet<PathBuf>, whitelisted_directories: BTreeSet<PathBuf>, ) -> Self

fn is_accessible(&self, path: &Path) -> bool

fn resolve(&self, path: &PathSegments) -> Option<PathBuf>

pub fn name(&self) -> &str

Returns the name of the volume

pub fn read_file(&self, path: &PathSegments) -> Option<OwnedBuffer>

Read a file from the volume.

pub fn read_dir( &self, path: &PathSegments, ) -> Option<Vec<(PathSegment, Option<[u8; 32]>, Metadata)>>

Read the contents of a directory.

pub fn metadata(&self, path: &PathSegments) -> Option<Metadata>

Get the metadata for a particular item.

pub(crate) fn as_directory_tree( &self, strictness: Strictness, ) -> Result<Directory<'_>, Error>