wasmer_package
Module utils

Functions

features_to_wasm_annotations

Convert a Features object to a list of WebAssembly feature strings that can be used in annotations.

from_bytes

from_disk

is_tarball ğŸ”’

Check if something looks like a *.tar.gz file.

parse_dir ğŸ”’

parse_tarball ğŸ”’

parse_v1_mmap ğŸ”’

parse_v2_mmap ğŸ”’

parse_v3_mmap ğŸ”’

wasm_annotations_to_features

Create a Features object from a list of WebAssembly feature strings.