wasmer_package

Module package

Load a Wasmer package from disk.

Modules

manifest 🔒

package 🔒

strictness 🔒

volume 🔒

Structs

FsVolume

A lazily loaded volume in a Wasmer package.

MemoryDir

An in-memory directory.

MemoryFile

An in-memory file.

MemoryVolume

An in-memory representation of a volume.

Package

A Wasmer package that will be lazily loaded from disk.

WalkBuilderFactory

Enums

ManifestError

Errors that may occur when converting from a wasmer_config::package::Manifest into a WebC manifest.

MemoryNode

An in-memory representation of a filesystem node.

Strictness

The strictness to use when working with a Package values from this crate.

WasmerPackageError

Errors that may occur while loading a Wasmer package from disk.

Traits

WasmerPackageVolume

An abstraction over concrete volumes implementation to be used in a Wasmer Package.

Functions

include_everything_walker

wasmer_ignore_walker

Create and configure a default ignore walker for building packages that: